refactor(carton): split profiler into cohesive modules
